pyproject.toml
[project]
name = "maf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework>=1.0.0b251120",
    "agent-framework-ag-ui>=1.0.0b251120",
    "agent-framework-azure-ai>=1.0.0b251120",
    "azure-ai-agents>=1.1.0",
    "azure-identity>=1.25.1",
    "fastapi>=0.121.3",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pre-commit>=4.5.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
